requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
django-heroku==0.3.1
django-import-export==1.2.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
PyYAML==5.1
sqlparse==0.3.0
tablib==0.13.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0